dm: omap: gpio: Put _get_gpio_value() logic into its own function
Add a separate internal helper function to get a GPIO value, so that we
will be able to call it with the driver model version and avoid code
duplication.

Also move gpio_get_bank() and check_gpio() down below the helper functions
as these won't be needed with driver model.

Change-Id: I06c0fdc39d231802ceebea006fe1790d032950f8
Signed-off-by: Simon Glass <sjg@chromium.org>