s5p: cpu_info: use defined CPU name if available
The CPU name for Exynos was concatenated with cpu id,
but for new Exynos platforms, like Chromebook Peach Pi
based on Exynos5800, the name of SoC variant does not
include the real SoC cpu id (0x5422).

For such case, the CPU name should be defined by the
config file.

This commit introduces new config:
- CONFIG_CPU_NAME - with cpu name string

If defined, then the cpu id is not printed.

Change-Id: If6166f349837e6b80a20dc932accdc42370cc5a4
Signed-off-by: Przemyslaw Marczak <p.marczak@samsung.com>